Upgrade pycairo version to 1.26.1 (#22110)
Why I did it
The build failed in 202311 due to an error while installing PyGObject 3.46.0, caused by a dependency conflict:

ERROR: Cannot install sonic-host-services==1.0 because these package versions have conflicting dependencies.

The conflict is caused by:
    sonic-host-services 1.0 depends on pycairo==1.26.1
    The user requested (constraint) pycairo==1.24.0
Our system includes a fallback mechanism that attempts to install the latest available package version if the specified version fails. However, the latest PyGObject release requires girepository-2.0, which is not supported in our environment.

To resolve this issue, this PR upgrades pycairo to 1.26.1, ensuring that the specified version of PyGObject 3.46.0 can be successfully installed.

How I did it
The build failed in 202311 due to an error while installing PyGObject 3.46.0, caused by a dependency conflict:

ERROR: Cannot install sonic-host-services==1.0 because these package versions have conflicting dependencies.

The conflict is caused by:
    sonic-host-services 1.0 depends on pycairo==1.26.1
    The user requested (constraint) pycairo==1.24.0
Our system includes a fallback mechanism that attempts to install the latest available package version if the specified version fails. However, the latest PyGObject release requires girepository-2.0, which is not supported in our environment.

To resolve this issue, this PR upgrades pycairo to 1.26.1, ensuring that the specified version of PyGObject 3.46.0 can be successfully installed.

How to verify it
VS build success in pipeline.