Switch from ntpd to chrony (#20497)
This replaces ntpd with Chrony, as described in sonic-net/SONiC#1852. The advantages of this (among others) is control over enabling/disabling long jumps/steps and guaranteed updates of the real time clock.

This PR also includes a submodule update of sonic-utilities, to bring in necessary changes there for chrony to work. The changelog for sonic-utilities is:

Submodule src/sonic-utilities ce51df2d8c..7cbb2f2fd5:
  > [sfputil] add support for sfputil debug tx-output/rx-output {port} enable/disable  (#3811)
  > Switch to using chrony instead of ntpd (#3574)
  > Added post commands for enabling fifos (#3801)
  > kdump-Remote-SSH-Configurations (#3400)

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>