Bump up /dev/shm size for Broadcom platforms to 512MB by default (#21772)
Why I did it
In the case of ASIC detection failures on Broadcom (or if the ASIC couldn't be detected in time), the /dev/shm partition in the syncd container will be only 64MB, which might cause issues if syncd/Broadcom SAI library needs more space than that.

How I did it
Since using a larger /dev/shm on its own doesn't cause any issues, bump up the default to 512MB. This should be enough for most platforms.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>