Enhance git checkout while building docker image.  (#22206)
Why I did it
We encountered the issue while checking out code.

fatal: reference is not a tree: xxx
##[warning]Git checkout failed on shallow repository, this might because of git fetch with depth '1' doesn't include the checkout commit 'xxx'. Please reference documentation (http://go.microsoft.com/fwlink/?LinkId=829603)
##[error]Git checkout failed with exit code: 128
In this PR, we address this issue by adding the checkout depth and ref branch.

Work item tracking
Microsoft ADO (number only):
How I did it
In this PR, we address this issue by adding the checkout depth and ref branch.