[docker-ptf]: Fix to set /run/sshd permission affecting 202311 (#22235)
Why I did it
The docker-ptf image built on 202311 release branch is created with a wrong permission for /run/sshd (/var/run/sshd) with value 0775. This fails to start sshd with the error below. Thus failing add-topo.

/run/sshd must be owned by root and not group or world-writable.
This PR fixes the permission to 0755 for /run/sshd (/var/run/sshd) in the docker-ptf image.

How I did it
Set permissions in sshd.conf supervisor config.

How to verify it
Local environment

Tested branch (Please provide the tested image version)
202311

Description for the changelog
[docker-ptf]: Fix to set /run/sshd permission affecting 202311

The docker-ptf image built on 202311 release branch is created with a wrong permission for /run/sshd (/var/run/sshd) with value 0775. This fails to start sshd with the error below. Thus failing add-topo.

/run/sshd must be owned by root and not group or world-writable.
This PR fixes the permission to 0755 for /run/sshd (/var/run/sshd) in the docker-ptf image.

Link to config_db schema for YANG module changes
NA