[Nokia] Lower the default warning threshold for fabric capacity (#21942)
Why I did it
MSFT has 7 SFM cards in Nokia chassis and each FAP will have 168 fabric links up with 7SFM cards. So 85% threshold is very high, so lowering the threshold to 75%

How I did it
Lowered the threshold to 75%

How to verify it
Verified in tested bed which has 8 IMMs and 7 SFM cards with fabric capacity sys logs.

Signed-off-by: saksarav <sakthivadivu.saravanaraj@nokia.com>