Set SAI_NUM_ECMP_MEMBERS for arista 7060x6_16pe_384c and 7060x6_b_64pe (#22221)
Why I did it
If we don't set the SAI_NUM_ECMP_MEMBERS attribute for SAI, the default value will be 64, that means we can't have any nexthop group with members more than 64, so we need to set the SAI_NUM_ECMP_MEMBERS explicitly.

How I did it
set the SAI_NUM_ECMP_MEMBERS explicitly.

How to verify it
Tested on testbed