[staticroutebfd]: Fix staticroutebfd's handling on blackhole static routes and static routes that only have partial info correctly (#22347)
Why I did it
Previously, the staticroutebfd module assume that static routes always have both ifname and nexthop configured, which is not true. First, a static route can have only either ifname or nexthop. Second, for blackhole routes, usually it is not configured with any ifname or nexthop. The staticroutebfd will encounter runtime error when those static routes are configured even if bfd is not enabled.

Fixes issue: #22344

How I did it
Change the staticroutebfd's implementation to skip the processing if the route is a blackhole route or it only has partial info. I also added UT to verify it.