Add a new switch_type called dummy-sup for voq chassis (#22393)
Supervisor on virtual VoQ chassis does not perform any functionality of a fabric card and does not have necessary config for being either a "voq" or "fabric" type of switch. Thus, we should mark it as a new type of switch object. I set the new type to be 'dummy-sup'. With this change, we should modify the criteria of determining whether a device is chassis.