Issue #22354: Support SET-REPLACE in gnxi/gnmi_cli_py (#22355)
By default, gnxi/gnmi_cli_py supports SET-REPLACE. It has been patched in to extend it to accept lists of paths. This patch did not preserve the REPLACE functionality. Fix the patch to keep the extended behavior but retain REPLACE handling.

Why I did it
gNMI is being extended to allow configs to be modified using GCU ConfigReplace. The syntax will be a JSON config passed as a file through gNMI using SET-REPLACE. sonic-mgmt tests use gnxi/gnmi_cli_py to test gNMI behavior, so in order to add tests for the new behavior gnxi/gnmi_cli_py must be support SET-REPLACE (see Issue#357).

Work item tracking
Microsoft ADO (number only):
How I did it
We can modify the offending patch directly. This seems better than adding a new patch just to rollback a previous one. It is also the last patch in the series, so there is no chance of downstream changes relying on it. There is already a set-type argument passed in to the _set() function which can be used to differentiate SET-REPLACE from SET-UPDATE.

How to verify it
sonic-mgmt/tests/gnmi/test_gnmi_configdb.py can be extended to test ConfigReplace by calling gnmi_set with a single config file as the replace_list. Without these change, this test will fail because the existing code will treat it as an update. With this change, the code sends the SET-REPLACE correctly and the test passes.