Added armed support for watchdog.py (#22643)
Why I did it
Watchdog is controlled by Pensando firmware container for Pensando DPU which is crucial for services running under docker container, hence it cannot be controlled. For watchdog-control.service (as per latest #22315), in case of dpu, it is going to be armed hence added support to check if watchdog is armed and what is timeout set for it.

Signed-off-by: Sahil Chaudhari <sahil.chaudhari@amd.com>