[Mellanox] Add a query in MDF to make copper-active modules be FW control (#22665)
- Why I did it
Nvidia doesn't support Copper Active modules to be SW control.
In case we use this kind of module, it should be classified as FW control.

- How I did it
Checked the media interface technology for the module. If it is Copper cable, linear active equalizers, it should be classified as FW control.

- How to verify it
Run this code on Mellanox switch with Copper cable, linear active equalizers module. check control sysfs to make sure it is classified as FW control. (control = 0).