[build]: Fix kdump build failure (#22553)
Fix #5097, Fix #17023

Why I did it
The build fails if kdump is enabled on the build host, even though the relevant build step is performed in a dockerized chroot.

+ sudo LANG=C chroot ./fsroot-cisco-8000 kdump-config symlinks 5.10.0-23-2-amd64
Cannot change symbolic links when kdump is loaded ... failed!

How I did it
kdump installation checks if kdump is already running and aborts if so. This is good in most cases, but it's not relevant when installing into a chroot inside a docker container. This adds a basic patch to disable this check during build.

Note that the kdump status of the build host is imported into the docker build container via the sysfs file system:

$ ls -id /sys/kernel/kexec_crash_loaded && cat /sys/kernel/kexec_crash_loaded
7824 /sys/kernel/kexec_crash_loaded
0

$ docker run --rm debian bash -c "ls -id /sys/kernel/kexec_crash_loaded && cat /sys/kernel/kexec_crash_loaded"
7824 /sys/kernel/kexec_crash_loaded
0
The inodes and file content are identical inside and outside of the container.

How to verify it
Enable kdump on the build host
Confirm baseline build fails with the "Cannot change symbolic links when kdump is loaded" error
Apply this change and build succeeds