[202505] Update the kernel to 6.1.123 (#22828)
* Update the kernel to 6.1.123 (#21711)

Why I did it
This updates the kernel to 6.1.123 to bring in bug and security fixes. This also includes a submodule update for sonic-linux-kernel and platform/broadcom/saibcm-modules-dnx.

sonic-linux-kernel:

aec83ae Update to Linux 6.1.123 (#454)
platform/broadom/saibcm-modules-dnx:

13a2b7a Update to Linux 6.1.123 (#22)
This also includes a hw-mgmt update for Mellanox platforms.

Work item tracking
Microsoft ADO (number only): 32198326

How to verify it
Verified that KVM and a physical device boots up, and verified that warm reboot of physical device works.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Co-authored-by: Vivek <vivekreddykarri98@gmail.com>

* Fix submodule hash to point to 202505 branch

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Co-authored-by: Vivek <vivekreddykarri98@gmail.com>