Network_Setup for Kdump Remote Mode. (#19464)
Why I did it
When the kdump remote feature is enabled by the user, a kernel crash results in an attempt to send the crash report file to an SSH server. However, the network interfaces are down, leading to a "Network Unreachable" error. To resolve this issue, I have added custom scripts to ensure the network interfaces are up and configured correctly during the kdump process.

How I did it
Added network_setup.sh and network_setup.hook file

How to verify it
After this change, the remote SSH kdump file transfer would be possible because a network not reachable issue arose as the default crash kernel image did not change the state of the interfaces to "UP" upon load.