YANG: Route Map allow match_interface on Loopback (#22889)
Why I did it
When using BGP Unnumbered for an underlay network its often desirable to only advertise the loopback interface IP address. The implementation allows it but the YANG model did not.

How I did it
Add yang model adjustments and adds a test to ensure it doesn't regress.

How to verify it
Run test case.

Description for the changelog
YANG: Route Map allow match_interface on Loopback