Enable impacted area based PR testing in sonic-buildimage repo (#22946)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log skip_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it
At the beginning of this year, we transitioned the PR testing approach in the sonic-mgmt repository from a classical model to an impacted area-based model. This new approach enables automatic collection of relevant test cases, eliminating the need for the previously hardcoded `pr_test_scripts.yaml` list, and we no longer maintain it.

However, the sonic-buildimage repository has continued to use the classical PR testing model. As a result, any new test cases added in the sonic-mgmt repository are not picked up during PR testing in sonic-buildimage.

This PR updates the sonic-buildimage repository to adopt the impacted area-based PR testing model. With this change, all relevant test cases—including newly added ones—will be properly included in the PR testing workflow.

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it
This PR updates the sonic-buildimage repository to adopt the impacted area-based PR testing model. With this change, all relevant test cases—including newly added ones—will be properly included in the PR testing workflow.

#### How to verify it
Test by pipeline itself, we can see that, the PR testing adopts the impacted area-based model and collect all relvent test cases.

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)