[dhcp_relay] Add clear dhcpv4 relay counter CLI (#22896)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log skip_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it
DHCPv4 per-interface counter is added by https://github.com/sonic-net/sonic-dhcpmon/pull/36. Current PR is to add CLI to clear counter.
HLD: https://github.com/sonic-net/SONiC/pull/1861

##### Work item tracking
- Microsoft ADO **(number only)**: 32751348

#### How I did it
1. Rename `dockers/docker-dhcp-relay/cli/clear/plugins/clear_dhcp6relay_counter.py` to `dockers/docker-dhcp-relay/cli/clear/plugins/clear_dhcp_relay.py` due to it's not only for dhcp6relay
2. Add support for clear DHCPv4 per-interface counter
3. Add UTs

#### How to verify it
1. Manually test by running `sudo sonic-clear dhcp_relay ipv4 counters`
2. UTs passed

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)