Network_Setup for Kdump Remote Mode. (#22999)
<!--
 Please make sure you've read and understood our contributing guidelines:
 https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

 failure_prs.log skip_prs.log Make sure all your commits include a signature generated with `git commit -s` **

 If this is a bug fix, make sure your description includes "fixes #xxxx", or
 "closes #xxxx" or "resolves #xxxx"

 Please provide the following information:
-->

#### Why I did it
When the kdump remote feature is enabled by the user, a kernel crash results in an attempt to send the crash report file to an SSH server. However, the network interfaces are down, leading to a "Network Unreachable" error. To resolve this issue, I have added custom scripts to ensure the network interfaces are up and configured correctly during the kdump process.

#### How I did it
Added network_setup.sh and network_setup.hook file

#### How to verify it
After this change, the remote SSH kdump file transfer would be possible because a network not reachable issue arose as the default crash kernel image did not change the state of the interfaces to "UP" upon load.
<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide a tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->