Add patch to fix multithreaded packet write issue in PTF (#21241)
Why I did it
This fixes a pcap write issue in the PTF framework, where if you have multiple threads using PTF, sending and receiving simultaneously, and PTF is generating a pcap file, then it's possible for the pcap file to end up being corrupted because two threads are writing to the pcap file at the same time.

This fix is cherry-picked from p4lang/ptf@e8b545f.

How I did it
Added a patch to the ptf-py3 build to fix this issue.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>