fix: temporarily disable sai validation (#23346)
Why I did it
Temporarily disable SAI validation for now as it is not compatible with the change in sonic-net/sonic-mgmt#19263 due to the usage of concurrent.futures. We will refactor the SAI validation and re-enable it later. Microsoft ADO to track the progress: 33758029.

Work item tracking
Microsoft ADO (number only): 33039693

signed-off-by: jianquanye@microsoft.com