[Arista]: Update Arista7280DR3A variants to ignore psu fans 3 and 4 (#23189)
These errors appear on Arista 72800DR3A SKUs. Pmon expects psu fan3 and fan4 to be readable but this particular model only has two psu fans, the rest need to be ignored.

Fixes #23188


How I did it
Modify sensors.conf to ignore these PSU fans.

How to verify it
Rebooted the DUT with the change and verified that the messages no longer appear.