Increase the installer disk size by 1GB and switch to ext2 for KVM (#23378)
Why I did it
Following #22199, the debug version of sonic-vs.bin file is at 4093MB (to be precise, one particular build was at 4291790831 bytes). This, in theory, should've fit fine in the 4096MB FAT32 installer disk that is created for building the sonic-vs.img QCOW2 image. However, it's short by 5MB. My guess for this reason is because FAT32 itself needs some space for metadata, and so we don't get the full 4096MB. This results in a failure to build the debug KVM image.

Work item tracking
Microsoft ADO (number only):
How I did it
As a fix, increase the installer disk size that we create by 1GB, which should be enough for now, and switch to ext2 filesystem instead of FAT32. This switch is needed because FAT32 only allows files up to 4GB (minus one byte), and the latest debug images are at around 4.1GB.

How to verify it
I locally made a loopback file of 4104MB, copied the sonic-vs.bin debug installer file generated in https://dev.azure.com/mssonic/build/_build/results?buildId=900622&view=results into it (which is 4093MB), and verified that it did get copied into there.