add interupts for alerts (#23377)
Why I did it
In SAI 12.x the SAI integrity counters is not supported completely. So, to detect packet drops due to packet corruptions or credit watchdog timeouts. We enable these interrupts and set the verbose level to error. This will generate syslogs which can be used to detect.

How I did it
Enable specific interrupt-ids via the sai_postinit_cmd.soc file.

How to verify it
Generate these interrupts manually and check if the syslog is generated

Signed-off-by: Arvindsrinivasan Lakshmi Narasimhan <arlakshm@microsoft.com>