[Reland] Update dockerfiles to do staged builds (#23164)
Why I did it
On newer versions of Docker, only the buildkit builder is supported, and
cannot be disabled by setting DOCKER_BUILDKIT to 0. The side effect of
this is that the behavior of --squash is different (see
moby/moby#38903). This will result in the container sizes being
significantly higher.

Work item tracking
Microsoft ADO (number only): 29277538
How I did it
To work around this, make all of our builds two-stage builds, with the
--squash flag entirely removed. The way this works is in the first
stage, whatever new files/packages need to be added are added (along
with files/packages that need to be removed). Then, in the second stage,
use rsync to copy over the changed files as a single command/layer. In the
case of the base layer for each Debian version, the final result of the first
stage will be copied into an empty base layer.

As part of this, also consolidate the container cleanup code into
post_run_cleanup, and remove it from the individual containers (for
consistency). Also experiment a bit with not needing to explicitly
install library dependencies, and let apt install it as necessary. This
will help during upgrades in the case of ABI changes for packages.

Also, remove the SONIC_USE_DOCKER_BUILDKIT option, and don't
set DOCKER_BUILDKIT option. This option will eventually have no impact.
This also means that builds will now use buildkit, as that is the default now.

How to verify it
Regular and debug VS image build at https://dev.azure.com/mssonic/build/_build/results?buildId=906196&view=results.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>