Blacklist amd_sfh for shearwater and redstart in aboot (#23511)
Why I did it
amd_sfh is a SOC that handles human interface devices using PCIe (ex: gyroscopic sensor, etc.). This is not used on the switch and can be blacklisted.

This module is throwing error messages at boot time due to the lack of devices to manage, where modprobe is failing to probe this device. This is failing mgmt tests (SRv6 tests). Blacklisting this device will resolve this issue.