Correctly fetch port config when HWSKU doesn't match between sonic-cfggen and configDB (#23185)
Why I did it
Incorrect golden config was being generated for non-default SKUs when using minigraph. This issue occurred because sonic-cfggen was reading port configuration from the config DB, even when the hardware SKU in the config DB differed. Incorrect ports and port_alias_map data from config db was causing the issue.

Work item tracking
Microsoft ADO (number only):
How I did it
Modified the logic to ensure that port configuration is fetched from the config DB only when the hardware SKU in the config DB matches the hardware SKU for which sonic-cfggen is generating the config.

How to verify it
Generated golden configurations for non-default SKUs and confirmed that the output was correct.