Don't attempt building the user container if the main slave fails (#23563)
Why I did it
If building the main slave container fails, then don't continue and try to build the user slave container (or try to collect the versions used in the failed slave container). This reduces the number of errors logged after the actual error message, and makes it a bit easier for reporting the actual cause of failures, besides saying that the user slave container failed to build because the main slave container wasn't available.

How I did it
Set the -o pipefail setting in the group command used to build the slave container, which will report the command as having failed if any command in a piped command fails. The bash shell already has -e set; however, in a piped command, the exit of the last command in the pipe is used in determining if the command failed. For example, with command1 | command2, without -o pipefail, the exit code of command1 is completely ignored. The docker build command is piped through | tee ... to save the output to a log file, which means -o pipefail needs to be set here.

How to verify it
Add a RUN false near the beginning of a slave container's Dockerfile to force a failure, and observe the behavior.