Update sai_stats_support_mask for Moby hwsku to remove PFC pause from port counter (#24062)
Why I did it
This PR is to set sai_stats_support_mask to 0x800 for Moby HWSKUs. This is a continuation of #23772

Work item tracking
Microsoft ADO 34515971:
CSP CS00012419125
How I did it
Update config.bcm

How to verify it
The change is verified by running PFCWD test on a physical testbed.