Add new HWSKU Arista-7060X6-16PE-384C-B-O128S4 (#24016)
Why I did it
As per the requeset in aristanetworks/sonic-qual.msft#844, this PR adds a new HWSKU Arista-7060X6-16PE-384C-B-O128S4 that supports 40G for 4x10G breakout on the QSFP28 port of Moby.

How I did it
Added Ethernet514 and Ethernet515 to anywhere needed, as well as add the corresponding logical and physical ports to the bcm config file.

How to verify it
Confirmed the links are up at 40G on our switch: