retry the chassis db cleanup operations (#24219)
When running load_minigraph or reloading configuration on the linecards, the interface-config.service restarts, which causes the midplane interface to flap. If swss.sh on the linecards deletes state from chassis_db, some states may not be cleaned up correctly, while others are successfully removed. For example, cleanup for SYSTEM_NEIGHBOR or SYSTEM_INTF may fail, but SYSTEM_LAG cleanup might succeed. This can lead to inconsistent lag IDs for the remote LC.

Signed-off-by: Arvindsrinivasan Lakshmi Narasimhan <arlakshm@microsoft.com>