Add support for use-route-replace in sonic dataplane provider (#23932)
ISSUE: https://github.com/sonic-net/sonic-buildimage/issues/23940

The FRR's zebra has two dataplane providers. One of them programs routes into the kernel, while the other one sends routes towards sonic. This one sends routes to fpmsyncd. Currently, this sonic dataplane provider does not make use of the route replace semantics. This means that any change to an existing route is first sent as a delete and then a new add notification is sent later. 
The sonic dataplane provider code currently sends a RTM_DELROUTE followed by a RTM_NEWROUTE for an update operation as well. This will lead to churn even when the ECMP set is growing.
This PR enables the "fpm use-route-replace" configuration in zebra. Additionally it also fixes the "show fpm status [json]" show cli.

Refer to the corresponding PR in fpmsyncd to be able to accept this https://github.com/sonic-net/sonic-swss/pull/3872

#### Why I did it

This will allow zebra to be configured with "fpm use-route-replace", which will cause it to update a route instead of delete followed by new.  This also helps when there are ECMP changes (ecmp growth or shrink) and allows for zero traffic drops. 

#### How I did it
Added support for route replace semantics in zebra's sonic dataplane provider.

#### How to verify it
Verified with ECMP routes with growth or shrink of ECMP.