yang: smart_switch dpu_id is defined as string
Passing an integer value in when the yang model is defined as a
string is invalid.  However, its possible the yang model itself
is not proper as it seems to limit the value to 1-7.  However,
such a change would be out of scope for this patch series as
we are just porting to libyang3.