[vpp]: Add sr0 interface for VPP. (#24359)
SONiC VPP has support for SRv6, but the sr0 interface is needed for it to work; added here in 202411: https://github.com/Azure/sonic-buildimage-msft/pull/517. A commit to master has comments requesting it to be enabled per platform: #24084. Adding this change so that for VPP it is enabled and can allow sonic-mgmt tests for srv6 to be run with Sonic-VPP.