Hardcode the hostname to use for syslogs (#24394)
* Hardcode the hostname to use for syslogs

On the host rsyslog, hardcode the hostname to use for logging syslogs,
and don't use the hostname that is sent from rsyslog in the container.
The reason for this is that when the container is started, the hostname
might not be set correctly (which is the case for the database
container); rsyslog reads the hostname once at process startup, but
doesn't reread it for each syslog message. This means that syslogs from
those containers might have the default hostname or an old hostname if
the rsyslogd process wasn't restarted after a hostname change.

To work around this, on the host rsyslog, when we write it to
`/var/log/syslog` or forward it out, use the hardcoded hostname that
gets set in the config_db file in place of rsyslog determining the
hostname or using the hostname passed down from the container rsyslog.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>