[202511][cherry-pick][FPM]: Cherry-pick Fix dplane_fpm_sonic for multiple SRv6 nexthops (#24839)
* [FPM]: Fix dplane_fpm_sonic for multiple SRv6 nexthops

Currently, there is a bug in the dplane_fpm_sonic to handle ECMP routes.
When FRR has an ECMP route with multiple SRv6 nexthops, the
dplane_fpm_sonic encodes the route with only the first SRv6 nextop in
the Netlink message, ignoring the other nexthops.

As a result, the Netlink message sent to fpmsyncd leads to the
installation of the route using only the first SRv6 nexthop, breaking
the EMCP nature of SRv6 routes.

This pull request fixes the issue in the dplane_fpm_sonic to handle all
SRv6 nexthops in the route.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>

* Fix the build failure (#24602)

Why I did it
Regression due to #23893

Work item tracking
Microsoft ADO (number only):
#24601
How I did it
Added the missing declaration of the referenced variable in the function fpm_nl_enqueue

How to verify it
Rebuilt the image, the compilation error was no longer seen.

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>

---------

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>
Co-authored-by: Vivek Verma <137406113+vivekverma-arista@users.noreply.github.com>